+++ airo-robots/airo_robots/grippers/parallel_position_gripper.py
