+++ airo-robots/airo_robots/manipulators/bimanual_position_manipulator.py
@@ -1,0 +1,4 @@
+class BimanualPositionManipulator:
+    """
+    (synchronous) base class for bimanual position-controlled manipulators. This could be a bimanual robot or a combination of 2 unimanual arms
+    """
